+++ building.cc
@@ -1,16 +1,16 @@
 #include "building.h"
 
-Building::Building(int index, string name, int purchaseCost) : Tile(index, name), purchaseCost(purchaseCost), owner(NULL), mortgaged(false) {}
+Building::Building(int index, string name, double purchaseCost) : Tile(index, name), purchaseCost(purchaseCost), owner(NULL), mortgaged(false) {}
 
 Player* Building::getOwner() {
     return owner;
 }
 
-int Building::getPurchaseCost() {
+double Building::getPurchaseCost() {
     return purchaseCost;
 }
 
-int Building::mortgage() {
+double Building::mortgage() {
 	mortgaged = true;
 	return purchaseCost / 2;
 }
